Rename: service -> application 패키지명 변경
